+++ src/onegov.feriennet/onegov/feriennet/collections/__init__.py
@@ -1,6 +1,5 @@
 from onegov.feriennet.collections.activity import VacationActivityCollection
 from onegov.feriennet.collections.billing import BillingCollection
-from onegov.feriennet.collections.billing import BillingDetails
 from onegov.feriennet.collections.match import MatchCollection
 
 from onegov.feriennet.collections.notification_template import (
@@ -13,7 +12,6 @@ from onegov.feriennet.collections.occasion_attendees import (
 
 __all__ = [
     'BillingCollection',
-    'BillingDetails',
     'MatchCollection',
     'NotificationTemplateCollection',
     'OccasionAttendeeCollection',