+++ src/onegov.feriennet/onegov/feriennet/custom.py
@@ -1,5 +1,5 @@
 from onegov.activity import BookingCollection
-from onegov.activity import InvoiceItemCollection
+from onegov.activity import InvoiceCollection
 from onegov.activity import PeriodCollection
 from onegov.feriennet import _, FeriennetApp
 from onegov.feriennet.collections import BillingCollection
@@ -101,10 +101,9 @@ def get_personal_tools(request):
         period = request.app.active_period
         periods = request.app.periods
 
-        invoice_items = InvoiceItemCollection(session, username)
-        unpaid = invoice_items.count_unpaid_invoices(
-            exclude_invoices={p.id.hex for p in periods if not p.finalized}
-        )
+        invoices = InvoiceCollection(session, user_id=request.current_user.id)
+        unpaid = invoices.unpaid_count(excluded_period_ids={
+            p.id for p in periods if not p.finalized})
 
         if unpaid:
             attributes = {
@@ -119,7 +118,7 @@ def get_personal_tools(request):
 
         yield Link(
             text=_("Invoices"),
-            url=request.link(invoice_items),
+            url=request.link(invoices),
             attrs=attributes
         )