+++ src/onegov.feriennet/onegov/feriennet/views/payment_overrides.py
@@ -1,4 +1,4 @@
-from onegov.activity import InvoiceItemCollection, InvoiceItem
+from onegov.activity import InvoiceCollection, InvoiceItem
 from onegov.core.security import Private
 from onegov.feriennet import FeriennetApp
 from onegov.pay import Payment, PaymentProviderCollection
@@ -16,13 +16,13 @@ def sync_payments_and_reconcile(self, request):
     # we keep two payment paid states, one in the invoice items table and one
     # in the payment (for online payments only).
     #
-    # when synching the payment states through the provider, we reconclie
+    # when syncing the payment states through the provider, we reconcile
     # those two states
     #
     # the reason we don't keep this all in the payment table is mainly a
     # performance one - we don't want to load the payments when rendering
     # the potentially large list of invoice items
-    InvoiceItemCollection(self.session).sync()
+    InvoiceCollection(self.session).sync()
 
     return result