+++ src/onegov.feriennet/onegov/feriennet/views/shared.py
@@ -2,7 +2,8 @@ from onegov.user import User, UserCollection
 from sqlalchemy import func
 
 
-def all_users(request):
+def users_for_select_element(request):
     u = UserCollection(request.session).query()
+    u = u.with_entities(User.id, User.username, User.title, User.realname)
     u = u.order_by(func.lower(User.title))
     return u.all()